Refactoring EventList component to be reused by all containers
