+++ App/components/EventList/styles.js
@@ -1,0 +1,16 @@
+export default StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#F5FCFF',
+  },
+  title: {
+    fontSize: 24,
+    textAlign: 'center',
+    margin: 15,
+  },
+  item: {
+    fontSize: 20,
+    textAlign: 'left',
+    margin: 13,
+  },
+});
