+++ App/views/HelperDashboard/index.js
@@ -5,8 +5,8 @@ import {
   Text,
   View
 } from 'react-native';
-import AssignedEventListContainer from '../../components/AssignedEventList/';
-import UnassignedEventListContainer from '../../components/UnassignedEventList/';
+import AssignedEventListContainer from '../../components/EventList/AssignedEventListContainer.js';
+import UnassignedEventListContainer from '../../components/EventList/UnassignedEventListContainer.js';
 
 export default class HelperDashboard extends Component {
   static navigationOptions = {
